:bug: float(x)

+++ config.py
@@ -71,7 +71,7 @@ PREDICT_TRANSFORM_Back = transforms.Compose([
                 transforms.Resize((101, 101)),
                 transforms.Grayscale(),
                 lambda x: x>0.5,
-                lambda x: x.astype(float)
+                lambda x: float(x)
             ])
